# _examples/northwind/api/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "northwind",
    "description": "Boilerplate code generated by **xo-grpc**. Modify `proto/typespb/v1/typespb.proto` file then run `buf generate` to change this description.",
    "version": "1.0",
    "contact": {
      "name": "xo-grpc",
      "url": "https://github.com/walterwanderley/xo-grpc"
    }
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/shipper": {
      "post": {
        "operationId": "ShipperService_Insert",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/shipperv1InsertRequest"
            }
          }
        ],
        "tags": [
          "ShipperService"
        ]
      }
    },
    "/v1/shipper/upsert": {
      "post": {
        "operationId": "ShipperService_Upsert",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/shipperv1UpsertRequest"
            }
          }
        ],
        "tags": [
          "ShipperService"
        ]
      }
    },
    "/v1/shipper/{shipperId}": {
      "get": {
        "operationId": "ShipperService_ShipperByShipperID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Shipper"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "shipperId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "ShipperService"
        ]
      },
      "delete": {
        "operationId": "ShipperService_Delete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "shipperId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "ShipperService"
        ]
      },
      "put": {
        "operationId": "ShipperService_Update",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "shipperId",
            "description": "Output only.",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "phone": {
                  "type": "string"
                },
                "shipperName": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "ShipperService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "shipperv1InsertRequest": {
      "type": "object",
      "properties": {
        "phone": {
          "type": "string"
        },
        "shipperId": {
          "type": "string",
          "format": "int64"
        },
        "shipperName": {
          "type": "string"
        }
      }
    },
    "shipperv1UpsertRequest": {
      "type": "object",
      "properties": {
        "phone": {
          "type": "string"
        },
        "shipperId": {
          "type": "string",
          "format": "int64"
        },
        "shipperName": {
          "type": "string"
        }
      }
    },
...
    "v1Shipper": {
      "type": "object",
      "properties": {
        "shipperId": {
          "type": "string",
          "format": "int64"
        },
        "shipperName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      }
    },
...
  }
}